[REF] change common.js to constants.js and add constants of command to constants.js
